refactor: Remove common `logMessage`
