+++ node/packages/aws-lambda-otel-extension/opt/otel-extension/external/helper.js
@@ -234,7 +234,14 @@ const measureAttributes = [
   },
 ];
 
+const logMessage = (...args) => {
+  if (process.env.DEBUG_SLS_OTEL_LAYER) {
+    console.log(...args);
+  }
+};
+
 module.exports = {
+  logMessage,
   resourceAttributes,
   measureAttributes,
   stripResponseBlobData,
