--- node/packages/aws-lambda-otel-extension/opt/otel-extension/lib/helper.js
@@ -1,11 +1,0 @@
-'use strict';
-
-const logMessage = (...args) => {
-  if (process.env.DEBUG_SLS_OTEL_LAYER) {
-    console.log(...args);
-  }
-};
-
-module.exports = {
-  logMessage,
-};
